Big update

+++ src/api/card.rs
@@ -2,7 +2,7 @@ use crate::models::{card::MemnixCard, mem::MemnixMem};
 
 type Result<T> = std::result::Result<T, Box<dyn std::error::Error + Send + Sync>>;
 
-/*
+
 pub async fn fetch_card(url: String) -> Result<MemnixCard> {
     let mut memnixcard = MemnixCard {
         question: "none".to_string(),
@@ -16,11 +16,11 @@ pub async fn fetch_card(url: String) -> Result<MemnixCard> {
         memnixcard = MemnixCard {
             question: echo_json["data"]["Card"]["card_question"].to_string(),
             answer: echo_json["data"]["Card"]["card_answer"].to_string(),
-            id: echo_json["data"]["Card"]["ID"].to_string().parse::<i8>().unwrap(),
+            id: echo_json["data"]["Card"]["ID"].to_string().parse::<u32>().unwrap(),
         };
     };
 
     Ok(memnixcard)
 }
-*/
+
